generate secure random number in Java
A secure way to generate a random number in Java with java.security.SecureRandom. SecureRandom uses random data from the os as its seed which makes it harder to predict than java.util.Random. The Random class uses the system clock as its seed which  can be predicted more easily.