Added guess_number.php
A simple PHP based guessing game, where user guesses a number from 1 to 10. The computer generates a random number from 1 to 10. If the number that the user enters matches the number that the computer generates, the user has guessed the correct number. If the number that the user has entered does not match the number that the computer has generated, the user is told that this is the incorrect number and to try again.